Finish the first two passes, which just regularize the tree.
It was simpler to work in two passes.  The first removes any 'nil' value for
'elements' and replaces it with the empty list.  The second replaces empty nodes
with their children and removes meaningless whitespace.

It's beginning to occur to me that it might be correct to translate this in to
our own AST which well suits our own needs rather than carrying Treetop's
representation too far forward.