Starting afresh with a different approach.  This can compile fixnums.
test-driver.rb is a full-loop test driver.  It takes test cases, parses them
(just using 'to_i' on this pass), passes them to the compiler, emits a prologue
and epilogue around the emitted source code, links everything with the C driver,
executes it, and verifies the output.

compiler.rb is the compiler.  It takes a parsed program as input and emits
assembly instructions.

driver.c is a very basic C-based driver.  It can print unencoded fixnums.

See also

 http://scheme2006.cs.uchicago.edu/11-ghuloum.pdf